refactor: restructure dashboard components
